+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Changelog
   | [#689](https://github.com/bugsnag/bugsnag-ruby/pull/689)
 * Add `time` (an ISO8601 string in UTC) to `device` metadata
   | [#690](https://github.com/bugsnag/bugsnag-ruby/pull/690)
+* Add `errors` to `Report`/`Event` containing an array of `Error` objects. The `Error` object contains `error_class`, `error_message` and `type` (always "ruby")
+  | [#691](https://github.com/bugsnag/bugsnag-ruby/pull/691)
 
 ### Fixes
 
@@ -37,6 +39,7 @@ Changelog
     * The `Breadcrumb#name` attribute has been deprecated in favour of `Breadcrumb#message`
     * The breadcrumb type constants in the `Bugsnag::Breadcrumbs` module has been deprecated in favour of the constants available in the `Bugsnag::BreadcrumbType` module
     For example, `Bugsnag::Breadcrumbs::ERROR_BREADCRUMB_TYPE` is now available as `Bugsnag::BreadcrumbType::ERROR`
+* `Report#exceptions` has been deprecated in favour of the new `errors` property
 
 ## v6.22.1 (11 August 2021)